MCU renderer: refactor the drawing code in its own module
